+++ CHANGES.md
@@ -1,4 +1,6 @@
 ## Unreleased
+
+- do not omit mandatory null Content-Length headers (mefyl #985)
 - cohttp-async, cohttp-curl-async: compatibility with core/async v0.16.0 (mseri, dkalinichenko-js #976)
 - cohttp-lwt server: call conn_closed before drainig the body of response on error (pirbo #982)
 - cohttp-eio: Relax socket interface requirement on `Server.connection_handler`. (mefyl #983)